Add files via upload
The 'most_common.py' script reads text from a file, tokenizes the text,  and normalizes the case. Then it creates a dictionary where he keys are individual the unique tokens in the text and the values are the frequencies of each token in the text.  Then we use Counter from collections to count the most frequent words in the text.